+++ src/types/context.h
@@ -31,6 +31,7 @@ Variable interface function template.
 #define une_variable_itf__(id__) une_variable *(id__)(une_context *context, wchar_t *name) 
 
 une_context *une_context_create(ptrdiff_t function);
+une_context *une_context_get_oldest_parent(une_context *context);
 void une_context_free_children(une_context *parent, une_context *youngest_child);
 void une_context_free(une_context *context);